Added the progress-element.
The progress element takes an optional max-attribute, which can be set to, for instance, 5 in an 5-step procedure. The 'min' value is always given as zero, no need to write it. If no max-attribute is given, the progress will go from 0.0 to 1.0. Therefore the max-attribute is optional in this snippet.

I've made the smippet exit within the element, so that you can write further snippets inside of it. An use case could be:

    <progress max="100"><span>0</span>%</progress>